More doc generation bug fixes
git-svn-id: https://svn.code.sf.net/p/ccpn/code/trunk@7650 8c04fef6-8de3-4ed0-b809-c157a10b9597

+++ python/ccpncore/memops/scripts/docgen/sphinxDocumentation.py
@@ -20,7 +20,8 @@ def refreshSphinxDocumentation():
   # Remove sphinx-apidoc files
   for ss in ('ccpn', 'ccpncore'):
     inDirectory = joinPath(docDirectory, 'source', ss)
-    shutil.rmtree(inDirectory)
+    if os.path.exists(inDirectory):
+      shutil.rmtree(inDirectory)
     os.mkdir(inDirectory)
 
   # clean builds
